requirements.txt
flask
flask-cors
ffmpeg-python
gunicorn